modified delete button and added new tests

+++ public/index.js
@@ -189,7 +189,7 @@ function renderAddNewEntryPage() {
             <input type="text" name="dishName" placeholder="Dish Name" id="dishName" class="editMeal" required>
             <label for="entryText">5. Review</label>
             <textarea name="entryText" cols="40" rows="5" id="entryText" placeholder="Detail your restaurant experience here" class="editMeal"></textarea>
-            <button class="deleteEntryButton">Add Entry</button>
+            <button class="submitEditButton">Add Entry</button>
         </form>
     `);
     handleAddEntryButton();
@@ -201,7 +201,7 @@ function handleUploadSubmitButton() {
     console.log('handleUploadSubmitButton ran');
     $('#uploadForm').on('click', '#photoUploadButton', (e) => {
         console.log('upload prevented default behavior')
-        var fd = new FormData($("form").get(0));
+        let fd = new FormData($("form").get(0));
         e.preventDefault();
         $.ajax({
             url: '/upload',
@@ -291,7 +291,6 @@ function handleRestaurantClick(restaurantsData) {
                     imgString = ``;
                 };
                 console.log(imgString);
-                //<img class="imgOffset" src="${meal.imageURL}" alt="${meal.dish}"/>
                 $('#mealList').append(`
                     <div class="card center">
                         <a class="card-link" href="#" id="${meal.id}">${meal.dish}</a>
@@ -390,20 +389,22 @@ function handleDeleteEntryButton(mealId, imageURL) {
             success: (res) => console.log(res),
             error: (err) => console.log(err)
         });
-        $.ajax({
-            url: '/upload',
-            type: 'DELETE',
-            headers: {
-                "Authorization": "Bearer " + localStorage.getItem('jwt')
-            },
-            contentType: "application/json; charset=utf-8",
-            dataType: "json",
-            data: JSON.stringify({
-                'imageURL': imageURL
-            }),
-            success: (res) => console.log(res),
-            error: (err) => console.log(err)
-        });
+        if (imageURL.length > 1) {
+            $.ajax({
+                url: '/upload',
+                type: 'DELETE',
+                headers: {
+                    "Authorization": "Bearer " + localStorage.getItem('jwt')
+                },
+                contentType: "application/json; charset=utf-8",
+                dataType: "json",
+                data: JSON.stringify({
+                    'imageURL': imageURL
+                }),
+                success: (res) => console.log(res),
+                error: (err) => console.log(err)
+            });
+        };
         getRestaurants();
     });
 }
